Add support for mapped memory files.

+++ lang8/src/main/java/net/openhft/chronicle/bytes/BytesStoreBytes.java
@@ -85,7 +85,7 @@ public class BytesStoreBytes implements Bytes {
     }
 
     @Override
-    public int refCount() {
+    public long refCount() {
         return bytesStore.refCount();
     }
